Revert merge commit until I get a couple things figured out..
~Need to figure out why I'm running into a copy-item issue for centbrowser_beta..

+++ automatic/axcrypt/tools/chocolateyinstall.ps1
@@ -5,6 +5,9 @@ $packageName = 'axcrypt'
 $toolsDir = Split-Path $MyInvocation.MyCommand.Definition
 $fileLocation = Get-Item "$toolsDir\*.exe"
 
+$registrypaths = @('HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\{9E15EF89-8322-C117-CAF2-E79EFAC71395}', 'HKLM:\SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall\{9E15EF89-8322-C117-CAF2-E79EFAC71395}')
+$version = '2.1.1541.0'
+
 $packageArgs = @{
   packageName    = $packageName
   fileType       = 'exe'
@@ -13,7 +16,22 @@ $packageArgs = @{
   validExitCodes = @(0)
 }
 
-Install-ChocolateyInstallPackage @packageArgs
+Foreach ($registry in $registrypaths) {
+  if (Test-Path $registry) {
+    $installedVersion = (
+      Get-ItemProperty -Path $registry -Name 'DisplayVersion'
+    ).DisplayVersion
+  }
+}
 
-# Remove the installers as there is no more need for it
-Remove-Item $toolsDir\*.exe -ea 0 -force
+if ($installedVersion -eq $version) {
+  Write-Output $(
+    "AxCrypt $installedVersion is already installed. " +
+    "Skipping download and installation."
+  )
+} else {
+  Install-ChocolateyInstallPackage @packageArgs
+
+  # Remove the installers as there is no more need for it
+  Remove-Item $toolsDir\*.exe -ea 0 -force
+}
